Renamed `content-card` and cleaned up Event Resources
Changed to content-card to match event-card for consistency
