+++ app/routes/messages/all-messages/components/all-messages.component.tsx
@@ -9,7 +9,7 @@ import {
 import { useMemo } from "react";
 
 import { SectionTitle } from "~/components";
-import { ContentCard } from "~/primitives/cards/content.card";
+import { ContentCard } from "~/primitives/cards/content-card";
 import { Icon } from "~/primitives/icon/icon";
 import { ContentItemHit } from "~/routes/search/types";
 import { CustomPagination } from "./custom-pagination.component";
